+++ Backend/src/main/java/dev/group21/rescmeal/repository/BusinessRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface BusinessRepository extends JpaRepository<Business, Integer> {
-    Optional<Business> findById(Integer id);
+public interface BusinessRepository extends JpaRepository<Business, Long> {
+    Optional<Business> findById(Long id);
 }
